{"schema":"decodex/commit/1","summary":"Modularize tracker records and account activity","authority":"manual"}
