+++ apps/decodex/src/agent/codex_accounts.rs
@@ -1,9 +1,6 @@
 use std::{
-	fs::{self, File, OpenOptions},
 	path::{Path, PathBuf},
-	process,
-	sync::{Mutex, OnceLock},
-	thread,
+	sync::Mutex,
 	time::Duration,
 };
 
@@ -12,15 +9,15 @@ use reqwest::{StatusCode, blocking::Client};
 use serde_json::Value;
 use time::{OffsetDateTime, format_description::well_known::Rfc3339};
 
-use crate::{
-	config::ProjectCodexAccountsConfig, prelude::eyre, runtime, state::CodexAccountActivitySummary,
-};
+use crate::{config::ProjectCodexAccountsConfig, prelude::eyre, runtime};
 
+mod activity;
 mod auth_failure;
 mod login;
 mod record;
 mod refresh;
 mod selection;
+mod storage;
 #[cfg(test)] mod tests;
 mod usage;
 
@@ -28,8 +25,7 @@ pub(crate) use self::{auth_failure::CodexAccountAuthFailure, login::CodexAccount
 use self::{
 	record::{
 		AccountPoolRecord, default_codex_auth_json_path,
-		default_profile_endpoint_for_usage_endpoint, parse_account_records,
-		sync_refreshed_record_to_codex_auth,
+		default_profile_endpoint_for_usage_endpoint, sync_refreshed_record_to_codex_auth,
 	},
 	refresh::{
 		ProactiveRefreshError, RefreshRequest, RefreshResponse, RefreshStatus,
@@ -38,14 +34,15 @@ use self::{
 	selection::{account_summaries, compare_account_candidates},
 	usage::{
 		AccountProfileSnapshot, AccountUsageSnapshot, UsageProbeError, nonblank_string,
-		preserve_cached_usage_windows, profile_snapshot_from_payload, usage_snapshot_from_payload,
+		profile_snapshot_from_payload, usage_snapshot_from_payload,
 	},
 };
 #[cfg(test)]
-pub(crate) use self::{
+pub(super) use self::{
 	refresh::{CodexTokenData, ProactiveRefreshReason},
 	usage::{CreditsSnapshot, UsageWindow},
 };
+#[cfg(test)] pub(super) use crate::state::CodexAccountActivitySummary;
 
 const DEFAULT_USAGE_ENDPOINT: &str = "https://chatgpt.com/backend-api/wham/usage";
 const DEFAULT_PROFILE_ENDPOINT: &str = "https://chatgpt.com/backend-api/wham/profiles/me";
@@ -54,10 +51,6 @@ const CODEX_USER_AGENT: &str = "codex-cli";
 const CHATGPT_OAUTH_CLIENT_ID: &str = "app_EMoamEEZ73f0CkXaXp7hrann";
 const HTTP_TIMEOUT: Duration = Duration::from_secs(10);
 const TOKEN_REFRESH_INTERVAL_SECONDS: i64 = 8 * 24 * 60 * 60;
-const ACCOUNT_ACTIVITY_CACHE_TTL_SECONDS: i64 = 60;
-const ACCOUNT_ACTIVITY_PROBE_MAX_CONCURRENCY: usize = 8;
-
-static ACCOUNT_ACTIVITY_CACHE: OnceLock<Mutex<Option<AccountActivityCacheEntry>>> = OnceLock::new();
 
 pub(crate) trait CodexAccountProvider {
 	fn select_account(&self) -> crate::prelude::Result<CodexAccountLogin>;
@@ -172,154 +165,6 @@ impl CodexAccountPool {
 			client,
 			selected_account_id: Mutex::new(None),
 		})
-	}
-
-	fn load_records(&self) -> crate::prelude::Result<Vec<AccountPoolRecord>> {
-		let input = fs::read_to_string(&self.path).map_err(|error| {
-			eyre::eyre!("Failed to read Codex accounts `{}`: {error}", self.path.display())
-		})?;
-
-		parse_account_records(&input, &self.path)
-	}
-
-	pub(crate) fn account_activity_summaries(
-		&self,
-	) -> crate::prelude::Result<Vec<CodexAccountActivitySummary>> {
-		let _guard = self.lock_records()?;
-		let mut records = self.load_records()?;
-
-		self.probe_account_activity_summaries(&mut records)
-	}
-
-	pub(crate) fn account_activity_summaries_cached(
-		&self,
-		force_refresh: bool,
-	) -> crate::prelude::Result<Vec<CodexAccountActivitySummary>> {
-		let now = OffsetDateTime::now_utc().unix_timestamp();
-		let cache_key = self.cache_key();
-		let cache = ACCOUNT_ACTIVITY_CACHE.get_or_init(|| Mutex::new(None));
-
-		if !force_refresh {
-			let cached = cache
-				.lock()
-				.map_err(|error| eyre::eyre!("Codex account usage cache is poisoned: {error}"))?;
-
-			if let Some(entry) = cached.as_ref()
-				&& entry.key == cache_key
-				&& now.saturating_sub(entry.checked_at_unix_epoch)
-					< ACCOUNT_ACTIVITY_CACHE_TTL_SECONDS
-			{
-				return Ok(entry.summaries.clone());
-			}
-		}
-
-		let mut summaries = self.account_activity_summaries()?;
-		let mut cached = cache
-			.lock()
-			.map_err(|error| eyre::eyre!("Codex account usage cache is poisoned: {error}"))?;
-
-		if let Some(entry) = cached.as_ref()
-			&& entry.key == cache_key
-		{
-			preserve_cached_usage_windows(&mut summaries, &entry.summaries, now);
-		}
-
-		*cached = Some(AccountActivityCacheEntry {
-			key: cache_key,
-			checked_at_unix_epoch: now,
-			summaries: summaries.clone(),
-		});
-
-		Ok(summaries)
-	}
-
-	pub(crate) fn account_activity_summaries_snapshot(
-		&self,
-	) -> crate::prelude::Result<Vec<CodexAccountActivitySummary>> {
-		let now = OffsetDateTime::now_utc().unix_timestamp();
-		let cache_key = self.cache_key();
-		let cache = ACCOUNT_ACTIVITY_CACHE.get_or_init(|| Mutex::new(None));
-		let cached = cache
-			.lock()
-			.map_err(|error| eyre::eyre!("Codex account usage cache is poisoned: {error}"))?;
-
-		if let Some(entry) = cached.as_ref()
-			&& entry.key == cache_key
-		{
-			return Ok(entry.summaries.clone());
-		}
-
-		drop(cached);
-
-		let _guard = self.lock_records()?;
-		let records = self.load_records()?;
-
-		Ok(records.iter().filter_map(|record| record.configured_activity_summary(now)).collect())
-	}
-
-	fn cache_key(&self) -> AccountActivityCacheKey {
-		AccountActivityCacheKey {
-			path: self.path.clone(),
-			usage_endpoint: self.usage_endpoint.clone(),
-			profile_endpoint: self.profile_endpoint.clone(),
-			refresh_endpoint: self.refresh_endpoint.clone(),
-		}
-	}
-
-	fn lock_records(&self) -> crate::prelude::Result<AccountPoolFileLock> {
-		let parent = self.path.parent().ok_or_else(|| {
-			eyre::eyre!(
-				"Codex accounts path `{}` must have a parent directory.",
-				self.path.display()
-			)
-		})?;
-		let file_name = self
-			.path
-			.file_name()
-			.and_then(|name| name.to_str())
-			.ok_or_else(|| eyre::eyre!("Codex accounts path must end in a valid file name."))?;
-		let lock_path = parent.join(format!(".{file_name}.lock"));
-		let file = OpenOptions::new()
-			.read(true)
-			.write(true)
-			.create(true)
-			.truncate(false)
-			.open(&lock_path)
-			.map_err(|error| {
-				eyre::eyre!("Failed to open Codex accounts lock `{}`: {error}", lock_path.display())
-			})?;
-
-		file.lock().map_err(|error| {
-			eyre::eyre!("Failed to lock Codex accounts `{}`: {error}", self.path.display())
-		})?;
-
-		Ok(AccountPoolFileLock { _file: file })
-	}
-
-	fn save_records(&self, records: &[AccountPoolRecord]) -> crate::prelude::Result<()> {
-		let parent = self.path.parent().ok_or_else(|| {
-			eyre::eyre!(
-				"Codex accounts path `{}` must have a parent directory.",
-				self.path.display()
-			)
-		})?;
-		let file_name = self
-			.path
-			.file_name()
-			.and_then(|name| name.to_str())
-			.ok_or_else(|| eyre::eyre!("Codex accounts path must end in a valid file name."))?;
-		let temp_path = parent.join(format!(".{file_name}.tmp-{}", process::id()));
-		let mut body = String::new();
-
-		for record in records {
-			body.push_str(&serde_json::to_string(record)?);
-			body.push('\n');
-		}
-
-		fs::write(&temp_path, body)?;
-		fs::rename(temp_path, &self.path)?;
-
-		Ok(())
 	}
 
 	fn select_from_records(
@@ -513,158 +358,6 @@ impl CodexAccountPool {
 			},
 			Err(error) => Ok(Err(format!("{} usage probe failed: {error}", record.display_name()))),
 		}
-	}
-
-	fn probe_account_activity_summaries(
-		&self,
-		records: &mut [AccountPoolRecord],
-	) -> crate::prelude::Result<Vec<CodexAccountActivitySummary>> {
-		let now = OffsetDateTime::now_utc().unix_timestamp();
-		let mut summaries_by_index = vec![None; records.len()];
-		let mut probe_inputs = Vec::new();
-		let mut records_changed = false;
-
-		for (index, record) in records.iter().enumerate() {
-			let Some(configured_summary) = record.configured_activity_summary(now) else {
-				continue;
-			};
-
-			if configured_summary.status != "available" {
-				summaries_by_index[index] = Some(configured_summary);
-
-				continue;
-			}
-
-			probe_inputs.push(AccountActivityProbeInput { index, record: record.clone() });
-		}
-		for chunk in probe_inputs.chunks(ACCOUNT_ACTIVITY_PROBE_MAX_CONCURRENCY) {
-			let results = self.probe_account_activity_summaries_parallel_chunk(chunk, now)?;
-
-			for result in results {
-				if result.records_changed {
-					records[result.index] = result.record;
-					records_changed = true;
-				}
-
-				summaries_by_index[result.index] = Some(result.summary);
-			}
-		}
-
-		if records_changed {
-			self.save_records(records)?;
-		}
-
-		Ok(summaries_by_index.into_iter().flatten().collect())
-	}
-
-	fn probe_account_activity_summaries_parallel_chunk(
-		&self,
-		inputs: &[AccountActivityProbeInput],
-		now: i64,
-	) -> crate::prelude::Result<Vec<AccountActivityProbeResult>> {
-		thread::scope(|scope| {
-			let handles = inputs
-				.iter()
-				.cloned()
-				.map(|input| scope.spawn(move || self.probe_account_activity_record(input, now)))
-				.collect::<Vec<_>>();
-			let mut results = Vec::with_capacity(handles.len());
-
-			for handle in handles {
-				let result = handle
-					.join()
-					.map_err(|_| eyre::eyre!("Codex account usage probe worker panicked."))??;
-
-				results.push(result);
-			}
-
-			Ok(results)
-		})
-	}
-
-	fn probe_account_activity_record(
-		&self,
-		input: AccountActivityProbeInput,
-		now: i64,
-	) -> crate::prelude::Result<AccountActivityProbeResult> {
-		let mut record = input.record;
-		let mut records_changed = false;
-		let refresh_status = match self.proactive_refresh_record(&mut record, now) {
-			Ok(status) => {
-				if status == RefreshStatus::Succeeded {
-					records_changed = true;
-				}
-
-				status.as_str()
-			},
-			Err(error) if error.auth_failed => {
-				records_changed = true;
-
-				let summary = record.auth_failed_activity_summary(now);
-
-				return Ok(AccountActivityProbeResult {
-					index: input.index,
-					record,
-					summary,
-					records_changed,
-				});
-			},
-			Err(error) if error.requires_skip => {
-				let summary = record.probe_failed_activity_summary(now, "failed", &error.source);
-
-				return Ok(AccountActivityProbeResult {
-					index: input.index,
-					record,
-					summary,
-					records_changed,
-				});
-			},
-			Err(_error) => "failed",
-		};
-		let summary = match self.probe_record_usage(&record) {
-			Ok(usage) => self.activity_summary_from_usage_probe(&record, usage, refresh_status)?,
-			Err(error) if error.unauthorized && record.refresh_token().is_some() => {
-				match self.refresh_record(&mut record) {
-					Ok(()) => {
-						records_changed = true;
-
-						match self.probe_record_usage(&record) {
-							Ok(usage) =>
-								self.activity_summary_from_usage_probe(&record, usage, "succeeded")?,
-							Err(retry_error) =>
-								record.probe_failed_activity_summary(now, "failed", &retry_error),
-						}
-					},
-					Err(refresh_error) => {
-						if refresh_error.downcast_ref::<CodexAccountAuthFailure>().is_some() {
-							records_changed = true;
-
-							record.auth_failed_activity_summary(now)
-						} else {
-							record.probe_failed_activity_summary(
-								now,
-								"failed",
-								refresh_error.as_ref(),
-							)
-						}
-					},
-				}
-			},
-			Err(error) => record.probe_failed_activity_summary(now, "probe_failed", &error),
-		};
-
-		Ok(AccountActivityProbeResult { index: input.index, record, summary, records_changed })
-	}
-
-	fn activity_summary_from_usage_probe(
-		&self,
-		record: &AccountPoolRecord,
-		usage: AccountUsageSnapshot,
-		refresh_status: &str,
-	) -> crate::prelude::Result<CodexAccountActivitySummary> {
-		let profile = self.probe_record_profile(record).ok().flatten();
-
-		record.activity_summary_from_usage_profile(usage, profile, refresh_status)
 	}
 
 	fn proactive_refresh_record(
@@ -954,36 +647,4 @@ impl CodexAccountProvider for CodexAccountPool {
 
 		self.refresh_from_records(&mut records, previous_account_id)
 	}
-}
-
-#[derive(Clone, Eq, PartialEq)]
-struct AccountActivityCacheKey {
-	path: PathBuf,
-	usage_endpoint: String,
-	profile_endpoint: Option<String>,
-	refresh_endpoint: String,
-}
-
-#[derive(Clone)]
-struct AccountActivityCacheEntry {
-	key: AccountActivityCacheKey,
-	checked_at_unix_epoch: i64,
-	summaries: Vec<CodexAccountActivitySummary>,
-}
-
-#[derive(Clone)]
-struct AccountActivityProbeInput {
-	index: usize,
-	record: AccountPoolRecord,
-}
-
-struct AccountActivityProbeResult {
-	index: usize,
-	record: AccountPoolRecord,
-	summary: CodexAccountActivitySummary,
-	records_changed: bool,
-}
-
-struct AccountPoolFileLock {
-	_file: File,
 }
