+++ apps/decodex/src/tracker/records.rs
@@ -1,975 +1,30 @@
-use std::path::{Component, Path};
+mod parse;
+mod public_projection;
+mod render;
+#[cfg(test)] mod tests;
+mod types;
+mod validation;
 
-use serde::{Deserialize, Serialize, de::DeserializeOwned};
-use serde_json::Error;
-
-use crate::tracker::{
-	TrackerComment,
-	privacy_classifier::{
-		PublicProjectionPrivacyClassification, PublicProjectionPrivacyClassifier,
-	},
-	public_text,
+#[cfg(test)]
+pub(crate) use self::public_projection::{
+	PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_ACTION, PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_COMMENT_BODY,
+	PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_DETAIL, PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_SUMMARY,
 };
-
 #[cfg(test)]
-pub(crate) const REVIEW_HANDOFF_RECORD_TYPE: &str = "review-handoff-record/1";
-#[cfg(test)]
-pub(crate) const CLOSEOUT_RECORD_TYPE: &str = "closeout-record/1";
-pub(crate) const LINEAR_EXECUTION_EVENT_RECORD_TYPE: &str = "decodex.linear_execution_event";
-pub(crate) const LINEAR_EXECUTION_EVENT_RECORD_VERSION: i64 = 1;
-pub(crate) const PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_SUMMARY: &str =
-	"Public summary withheld by local privacy classifier.";
-
-const PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_ACTION: &str =
-	"Public action withheld by local privacy classifier; review private Decodex evidence.";
-const PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_DETAIL: &str =
-	"Public detail withheld by local privacy classifier.";
-const PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_COMMENT_BODY: &str =
-	"Public comment details withheld by local privacy classifier.";
-
-#[cfg(test)]
-#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
-pub(crate) struct ReviewHandoffRecord {
-	#[serde(rename = "type")]
-	pub(crate) record_type: String,
-	pub(crate) completed_at: String,
-	pub(crate) run_id: String,
-	pub(crate) attempt_number: i64,
-	pub(crate) branch_name: String,
-	pub(crate) pr_url: String,
-	pub(crate) target_base_ref_name: String,
-	pub(crate) pr_head_ref_name: String,
-	pub(crate) pr_head_oid: String,
-	pub(crate) summary: String,
-}
-
-#[cfg(test)]
-#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
-pub(crate) struct CloseoutRecord {
-	#[serde(rename = "type")]
-	pub(crate) record_type: String,
-	pub(crate) completed_at: String,
-	pub(crate) run_id: String,
-	pub(crate) attempt_number: i64,
-	pub(crate) branch_name: String,
-	pub(crate) pr_url: String,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
-#[serde(deny_unknown_fields)]
-pub(crate) struct LinearExecutionEventRecord {
-	pub(crate) record_type: String,
-	pub(crate) record_version: i64,
-	pub(crate) event_type: String,
-	pub(crate) event_timestamp: String,
-	pub(crate) idempotency_key: String,
-	pub(crate) service_id: String,
-	pub(crate) issue_id: String,
-	pub(crate) issue_identifier: String,
-	pub(crate) run_id: String,
-	pub(crate) attempt_number: i64,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) branch: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) worktree_path: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) commit_sha: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) pr_url: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) pr_head_sha: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) pr_base_ref: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) summary: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) validation_result: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) phase: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) focus: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) next_action: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) blockers: Option<Vec<String>>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) evidence: Option<Vec<String>>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) verification: Option<Vec<String>>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) error_class: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) terminal_path: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) cleanup_status: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) transport: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) target_state: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) failed_command: Option<String>,
-	#[serde(skip_serializing_if = "Option::is_none")]
-	pub(crate) raw_error: Option<String>,
-}
-impl LinearExecutionEventRecord {
-	pub(crate) fn new(
-		identity: LinearExecutionEventIdentity<'_>,
-		event_type: &str,
-		event_timestamp: String,
-		stable_anchor: &str,
-	) -> Self {
-		Self {
-			record_type: String::from(LINEAR_EXECUTION_EVENT_RECORD_TYPE),
-			record_version: LINEAR_EXECUTION_EVENT_RECORD_VERSION,
-			event_type: event_type.to_owned(),
-			event_timestamp,
-			idempotency_key: linear_execution_idempotency_key(
-				identity.service_id,
-				identity.issue_identifier,
-				identity.run_id,
-				identity.attempt_number,
-				event_type,
-				stable_anchor,
-			),
-			service_id: identity.service_id.to_owned(),
-			issue_id: identity.issue_id.to_owned(),
-			issue_identifier: identity.issue_identifier.to_owned(),
-			run_id: identity.run_id.to_owned(),
-			attempt_number: identity.attempt_number,
-			branch: None,
-			worktree_path: None,
-			commit_sha: None,
-			pr_url: None,
-			pr_head_sha: None,
-			pr_base_ref: None,
-			summary: None,
-			validation_result: None,
-			phase: None,
-			focus: None,
-			next_action: None,
-			blockers: None,
-			evidence: None,
-			verification: None,
-			error_class: None,
-			terminal_path: None,
-			cleanup_status: None,
-			transport: None,
-			target_state: None,
-			failed_command: None,
-			raw_error: None,
-		}
-	}
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) struct LinearExecutionEventIdentity<'a> {
-	pub(crate) service_id: &'a str,
-	pub(crate) issue_id: &'a str,
-	pub(crate) issue_identifier: &'a str,
-	pub(crate) run_id: &'a str,
-	pub(crate) attempt_number: i64,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) struct LinearExecutionEventPublicProjection {
-	pub(crate) body: String,
-	pub(crate) record: LinearExecutionEventRecord,
-	pub(crate) classifier_withheld_text: bool,
-}
-
-pub(crate) fn render_linear_execution_event_comment_body(
-	record: &LinearExecutionEventRecord,
-	retry_budget_attempt_count: Option<i64>,
-) -> String {
-	let mut body = format!(
-		"Decodex execution event: {}\n\n- run_sequence_attempt: `{}` (not retry-budget count)",
-		record.event_type, record.attempt_number
-	);
-
-	if let Some(retry_budget_attempt_count) = retry_budget_attempt_count {
-		body.push_str(&format!(
-			"\n- retry_budget_attempts_consumed: `{retry_budget_attempt_count}`"
-		));
-	}
-
-	body
-}
-
-pub(crate) fn linear_execution_event_public_projection(
-	body: &str,
-	record: &LinearExecutionEventRecord,
-	privacy_classifier: &dyn PublicProjectionPrivacyClassifier,
-) -> LinearExecutionEventPublicProjection {
-	let (body, body_withheld) = classify_public_comment_body(body, privacy_classifier);
-	let (record, record_withheld) =
-		classify_linear_execution_event_record(record, privacy_classifier);
-
-	LinearExecutionEventPublicProjection {
-		body,
-		record,
-		classifier_withheld_text: body_withheld || record_withheld,
-	}
-}
-
-/// Render the low-frequency public Linear projection for a private progress checkpoint.
-pub(crate) fn render_progress_checkpoint_public_projection(
-	identity: LinearExecutionEventIdentity<'_>,
-	event_timestamp: String,
-	phase: &str,
-	branch: Option<&str>,
-	worktree_path: Option<&str>,
-	pr_url: Option<&str>,
-) -> LinearExecutionEventRecord {
-	let anchor = stable_event_anchor(&[
-		phase,
-		branch.unwrap_or_default(),
-		worktree_path.unwrap_or_default(),
-		pr_url.unwrap_or_default(),
-	]);
-	let mut record =
-		LinearExecutionEventRecord::new(identity, "progress_checkpoint", event_timestamp, &anchor);
-
-	record.phase = Some(phase.to_owned());
-	record.branch = branch.map(ToOwned::to_owned);
-	record.worktree_path = worktree_path.map(ToOwned::to_owned);
-	record.pr_url = pr_url.map(ToOwned::to_owned);
-	record.summary = Some(format!("Execution phase: {phase}."));
-
-	record
-}
-
-pub(crate) fn format_structured_comment<T>(record: &T) -> std::result::Result<String, Error>
-where
-	T: Serialize,
-{
-	let payload = serde_json::to_string_pretty(record)?;
-
-	Ok(format!("```json\n{payload}\n```"))
-}
-
-pub(crate) fn append_structured_comment_record<T>(
-	body: &str,
-	record: &T,
-) -> std::result::Result<String, Error>
-where
-	T: Serialize,
-{
-	let payload = format_structured_comment(record)?;
-
-	if body.trim().is_empty() {
-		return Ok(payload);
-	}
-
-	Ok(format!("{body}\n\n{payload}"))
-}
-
-pub(crate) fn stable_event_anchor(parts: &[&str]) -> String {
-	let mut hash = 0xcbf29ce484222325_u64;
-
-	for part in parts {
-		for byte in part.as_bytes() {
-			hash ^= u64::from(*byte);
-			hash = hash.wrapping_mul(0x100000001b3);
-		}
-
-		hash ^= 0xff;
-		hash = hash.wrapping_mul(0x100000001b3);
-	}
-
-	format!("{hash:016x}")
-}
-
-pub(crate) fn validate_linear_execution_event_record(
-	record: &LinearExecutionEventRecord,
-) -> Result<(), String> {
-	validate_linear_execution_event_envelope(record)?;
-	validate_linear_execution_event_fields(record)?;
-	validate_linear_execution_event_public_text(record)?;
-
-	if let Some(worktree_path) = record.worktree_path.as_deref() {
-		validate_repo_relative_path(worktree_path, "worktree_path")?;
-	}
-
-	Ok(())
-}
-
-pub(crate) fn has_linear_execution_event_record(
-	comments: &[TrackerComment],
-	service_id: &str,
-	issue_id: &str,
-	idempotency_key: &str,
-) -> bool {
-	comments.iter().filter_map(|comment| parse_linear_execution_event_record(&comment.body)).any(
-		|record| {
-			record.service_id == service_id
-				&& record.issue_id == issue_id
-				&& record.idempotency_key == idempotency_key
-		},
-	)
-}
-
-pub(crate) fn parse_linear_execution_event_record(
-	body: &str,
-) -> Option<LinearExecutionEventRecord> {
-	parse_structured_comment::<LinearExecutionEventRecord>(body)
-		.filter(|record| validate_linear_execution_event_record(record).is_ok())
-}
-
-fn validate_linear_execution_event_public_text(
-	record: &LinearExecutionEventRecord,
-) -> Result<(), String> {
-	for (field_name, value) in [
-		("summary", record.summary.as_deref()),
-		("focus", record.focus.as_deref()),
-		("next_action", record.next_action.as_deref()),
-		("failed_command", record.failed_command.as_deref()),
-		("raw_error", record.raw_error.as_deref()),
-	] {
-		if let Some(value) = value {
-			public_text::validate_public_text_field(field_name, value)?;
-		}
-	}
-	for (field_name, values) in [
-		("blockers", record.blockers.as_ref()),
-		("evidence", record.evidence.as_ref()),
-		("verification", record.verification.as_ref()),
-	] {
-		if let Some(values) = values {
-			public_text::validate_public_text_items(field_name, values)?;
-		}
-	}
-
-	Ok(())
-}
-
-fn classify_public_comment_body(
-	body: &str,
-	privacy_classifier: &dyn PublicProjectionPrivacyClassifier,
-) -> (String, bool) {
-	if body.trim().is_empty() {
-		return (String::new(), false);
-	}
-	if classifier_allows(privacy_classifier, "body", body) {
-		return (body.to_owned(), false);
-	}
-
-	(String::from(PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_COMMENT_BODY), true)
-}
-
-fn classify_linear_execution_event_record(
-	record: &LinearExecutionEventRecord,
-	privacy_classifier: &dyn PublicProjectionPrivacyClassifier,
-) -> (LinearExecutionEventRecord, bool) {
-	let mut record = record.clone();
-	let event_type = record.event_type.clone();
-	let mut withheld = false;
-
-	classify_optional_text_field(
-		&mut record.summary,
-		"summary",
-		event_requires_summary(&event_type),
-		PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_SUMMARY,
-		privacy_classifier,
-		&mut withheld,
-	);
-	classify_optional_text_field(
-		&mut record.focus,
-		"focus",
-		false,
-		PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_DETAIL,
-		privacy_classifier,
-		&mut withheld,
-	);
-	classify_optional_text_field(
-		&mut record.next_action,
-		"next_action",
-		event_requires_next_action(&event_type),
-		PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_ACTION,
-		privacy_classifier,
-		&mut withheld,
-	);
-	classify_optional_text_field(
-		&mut record.failed_command,
-		"failed_command",
-		false,
-		PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_DETAIL,
-		privacy_classifier,
-		&mut withheld,
-	);
-	classify_optional_text_field(
-		&mut record.raw_error,
-		"raw_error",
-		false,
-		PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_DETAIL,
-		privacy_classifier,
-		&mut withheld,
-	);
-	classify_optional_text_items(
-		&mut record.blockers,
-		"blockers",
-		event_requires_items(&event_type, "blockers"),
-		privacy_classifier,
-		&mut withheld,
-	);
-	classify_optional_text_items(
-		&mut record.evidence,
-		"evidence",
-		event_requires_items(&event_type, "evidence"),
-		privacy_classifier,
-		&mut withheld,
-	);
-	classify_optional_text_items(
-		&mut record.verification,
-		"verification",
-		false,
-		privacy_classifier,
-		&mut withheld,
-	);
-
-	(record, withheld)
-}
-
-fn classify_optional_text_field(
-	value: &mut Option<String>,
-	field_name: &str,
-	required: bool,
-	fallback: &str,
-	privacy_classifier: &dyn PublicProjectionPrivacyClassifier,
-	withheld: &mut bool,
-) {
-	let Some(current) = value.as_deref() else {
-		return;
-	};
-
-	if classifier_allows(privacy_classifier, field_name, current) {
-		return;
-	}
-
-	*withheld = true;
-
-	if required {
-		*value = Some(fallback.to_owned());
-	} else {
-		*value = None;
-	}
-}
-
-fn classify_optional_text_items(
-	values: &mut Option<Vec<String>>,
-	field_name: &str,
-	required: bool,
-	privacy_classifier: &dyn PublicProjectionPrivacyClassifier,
-	withheld: &mut bool,
-) {
-	let Some(current_values) = values.take() else {
-		return;
-	};
-	let mut retained = Vec::new();
-
-	for value in current_values {
-		if classifier_allows(privacy_classifier, field_name, &value) {
-			retained.push(value);
-		} else {
-			*withheld = true;
-		}
-	}
-
-	if retained.is_empty() {
-		*values = required.then(|| vec![String::from(PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_DETAIL)]);
-	} else {
-		*values = Some(retained);
-	}
-}
-
-fn classifier_allows(
-	privacy_classifier: &dyn PublicProjectionPrivacyClassifier,
-	field_name: &str,
-	value: &str,
-) -> bool {
-	matches!(
-		privacy_classifier.classify_public_projection_text(field_name, value),
-		PublicProjectionPrivacyClassification::Allow
-	)
-}
-
-fn event_requires_summary(event_type: &str) -> bool {
-	matches!(
-		event_type,
-		"run_started"
-			| "intake"
-			| "progress_checkpoint"
-			| "review_handoff"
-			| "repair_handoff"
-			| "review_handoff_rebind"
-			| "review_handoff_adopt"
-			| "landed"
-			| "closeout"
-			| "cleanup_complete"
-	)
-}
-
-fn event_requires_next_action(event_type: &str) -> bool {
-	matches!(event_type, "needs_attention" | "terminal_failure")
-}
-
-fn event_requires_items(event_type: &str, field_name: &str) -> bool {
-	matches!(
-		(event_type, field_name),
-		(
-			"needs_attention"
-				| "terminal_failure"
-				| "review_handoff_rebind"
-				| "review_handoff_adopt",
-			"evidence",
-		) | ("needs_attention" | "terminal_failure", "blockers")
-	)
-}
-
-fn parse_structured_comment<T>(body: &str) -> Option<T>
-where
-	T: DeserializeOwned,
-{
-	extract_structured_json_blocks(body)
-		.into_iter()
-		.rev()
-		.find_map(|payload| serde_json::from_str::<T>(payload).ok())
-		.or_else(|| serde_json::from_str::<T>(body.trim()).ok())
-}
-
-fn extract_structured_json_blocks(body: &str) -> Vec<&str> {
-	body.match_indices("```json")
-		.filter_map(|(start, _)| {
-			let fenced = &body[start + "```json".len()..];
-			let fenced = fenced.strip_prefix("\r\n").or_else(|| fenced.strip_prefix('\n'))?;
-			let end = fenced.find("\n```").or_else(|| fenced.find("\r\n```"))?;
-
-			Some(fenced[..end].trim())
-		})
-		.collect()
-}
-
-fn linear_execution_idempotency_key(
-	service_id: &str,
-	issue_identifier: &str,
-	run_id: &str,
-	attempt_number: i64,
-	event_type: &str,
-	stable_anchor: &str,
-) -> String {
-	format!(
-		"{service_id}:{issue_identifier}:{run_id}:{attempt_number}:{event_type}:{stable_anchor}"
-	)
-}
-
-fn validate_linear_execution_event_envelope(
-	record: &LinearExecutionEventRecord,
-) -> Result<(), String> {
-	if record.record_type != LINEAR_EXECUTION_EVENT_RECORD_TYPE {
-		return Err(format!("`record_type` must be `{LINEAR_EXECUTION_EVENT_RECORD_TYPE}`."));
-	}
-	if record.record_version != LINEAR_EXECUTION_EVENT_RECORD_VERSION {
-		return Err(format!("`record_version` must be `{LINEAR_EXECUTION_EVENT_RECORD_VERSION}`."));
-	}
-
-	for (field, value) in [
-		("event_type", record.event_type.as_str()),
-		("event_timestamp", record.event_timestamp.as_str()),
-		("idempotency_key", record.idempotency_key.as_str()),
-		("service_id", record.service_id.as_str()),
-		("issue_id", record.issue_id.as_str()),
-		("issue_identifier", record.issue_identifier.as_str()),
-		("run_id", record.run_id.as_str()),
-	] {
-		if value.trim().is_empty() {
-			return Err(format!("`{field}` must not be empty."));
-		}
-	}
-
-	if record.attempt_number < 1 {
-		return Err(String::from("`attempt_number` must be at least 1."));
-	}
-
-	Ok(())
-}
-
-fn validate_linear_execution_event_fields(
-	record: &LinearExecutionEventRecord,
-) -> Result<(), String> {
-	match record.event_type.as_str() {
-		"run_started" => {
-			require_string(record.branch.as_deref(), "branch")?;
-			require_string(record.worktree_path.as_deref(), "worktree_path")?;
-			require_string(record.commit_sha.as_deref(), "commit_sha")?;
-			require_string(record.transport.as_deref(), "transport")?;
-
-			require_string(record.summary.as_deref(), "summary")
-		},
-		"intake" => require_string(record.summary.as_deref(), "summary"),
-		"lease_acquired" => {
-			require_string(record.branch.as_deref(), "branch")?;
-
-			Ok(())
-		},
-		"worktree_prepared" => {
-			require_string(record.branch.as_deref(), "branch")?;
-			require_string(record.worktree_path.as_deref(), "worktree_path")?;
-
-			require_string(record.commit_sha.as_deref(), "commit_sha")
-		},
-		"agent_started" => {
-			require_string(record.branch.as_deref(), "branch")?;
-
-			require_string(record.worktree_path.as_deref(), "worktree_path")
-		},
-		"progress_checkpoint" => {
-			require_string(record.phase.as_deref(), "phase")?;
-			require_string(record.summary.as_deref(), "summary")?;
-
-			reject_progress_checkpoint_private_fields(record)
-		},
-		"pr_opened" | "pr_updated" => validate_pr_event_fields(record),
-		"review_handoff" | "repair_handoff" => {
-			validate_pr_event_fields(record)?;
-			require_string(record.validation_result.as_deref(), "validation_result")?;
-			require_string(record.summary.as_deref(), "summary")?;
-
-			require_string(record.terminal_path.as_deref(), "terminal_path")
-		},
-		"review_handoff_rebind" | "review_handoff_adopt" => {
-			validate_pr_event_fields(record)?;
-			require_string(record.validation_result.as_deref(), "validation_result")?;
-			require_string(record.summary.as_deref(), "summary")?;
-
-			require_vec(record.evidence.as_ref(), "evidence")
-		},
-		"landed" => {
-			validate_pr_event_fields(record)?;
-
-			require_string(record.summary.as_deref(), "summary")
-		},
-		"closeout" => {
-			require_string(record.pr_url.as_deref(), "pr_url")?;
-			require_string(record.commit_sha.as_deref(), "commit_sha")?;
-
-			require_string(record.summary.as_deref(), "summary")
-		},
-		"needs_attention" => {
-			require_string(record.error_class.as_deref(), "error_class")?;
-			require_string(record.next_action.as_deref(), "next_action")?;
-			require_vec(record.blockers.as_ref(), "blockers")?;
-			require_vec(record.evidence.as_ref(), "evidence")?;
-
-			require_string(record.terminal_path.as_deref(), "terminal_path")
-		},
-		"terminal_failure" => {
-			require_string(record.error_class.as_deref(), "error_class")?;
-			require_string(record.next_action.as_deref(), "next_action")?;
-			require_vec(record.blockers.as_ref(), "blockers")?;
-
-			require_vec(record.evidence.as_ref(), "evidence")
-		},
-		"cleanup_complete" => {
-			require_string(record.branch.as_deref(), "branch")?;
-			require_string(record.worktree_path.as_deref(), "worktree_path")?;
-			require_string(record.cleanup_status.as_deref(), "cleanup_status")?;
-
-			require_string(record.summary.as_deref(), "summary")
-		},
-		other => Err(format!("Unsupported Linear execution event type `{other}`.")),
-	}
-}
-
-fn reject_progress_checkpoint_private_fields(
-	record: &LinearExecutionEventRecord,
-) -> Result<(), String> {
-	for field in [
-		("focus", record.focus.is_some()),
-		("next_action", record.next_action.is_some()),
-		("blockers", record.blockers.is_some()),
-		("evidence", record.evidence.is_some()),
-		("verification", record.verification.is_some()),
-		("commit_sha", record.commit_sha.is_some()),
-	] {
-		let (field_name, present) = field;
-
-		if present {
-			return Err(format!(
-				"`progress_checkpoint` Linear events must use the public projection; `{field_name}` belongs in private execution events."
-			));
-		}
-	}
-
-	Ok(())
-}
-
-fn validate_pr_event_fields(record: &LinearExecutionEventRecord) -> Result<(), String> {
-	require_string(record.branch.as_deref(), "branch")?;
-	require_string(record.pr_url.as_deref(), "pr_url")?;
-	require_string(record.pr_head_sha.as_deref(), "pr_head_sha")?;
-	require_string(record.pr_base_ref.as_deref(), "pr_base_ref")?;
-
-	require_string(record.commit_sha.as_deref(), "commit_sha")
-}
-
-fn require_string(value: Option<&str>, field: &str) -> Result<(), String> {
-	if value.is_some_and(|value| !value.trim().is_empty()) {
-		return Ok(());
-	}
-
-	Err(format!("`{field}` is required for this Linear execution event."))
-}
-
-fn require_vec(value: Option<&Vec<String>>, field: &str) -> Result<(), String> {
-	if value.is_some() {
-		return Ok(());
-	}
-
-	Err(format!("`{field}` is required for this Linear execution event."))
-}
-
-fn validate_repo_relative_path(path: &str, field_name: &str) -> Result<(), String> {
-	if path.is_empty() {
-		return Err(format!("`{field_name}` must not be empty."));
-	}
-	if path.starts_with('/') || path.starts_with("~/") || has_drive_root_prefix(path) {
-		return Err(format!("`{field_name}` must be repository-relative, not `{path}`."));
-	}
-	if Path::new(path).components().any(|component| matches!(component, Component::ParentDir)) {
-		return Err(format!("`{field_name}` must stay within the repository, not `{path}`."));
-	}
-
-	Ok(())
-}
-
-fn has_drive_root_prefix(path: &str) -> bool {
-	let bytes = path.as_bytes();
-
-	bytes.len() >= 3
-		&& bytes[0].is_ascii_alphabetic()
-		&& bytes[1] == b':'
-		&& matches!(bytes[2], b'\\' | b'/')
-}
-
-#[cfg(test)]
-mod tests {
-	use crate::tracker::{
-		privacy_classifier::{
-			PublicProjectionPrivacyClassification, PublicProjectionPrivacyClassifier,
-		},
-		records::{LinearExecutionEventIdentity, LinearExecutionEventRecord},
-	};
-
-	struct AllowingClassifier;
-	impl PublicProjectionPrivacyClassifier for AllowingClassifier {
-		fn classify_public_projection_text(
-			&self,
-			_field_name: &str,
-			_text: &str,
-		) -> PublicProjectionPrivacyClassification {
-			PublicProjectionPrivacyClassification::Allow
-		}
-	}
-
-	struct SuspiciousWordClassifier;
-	impl PublicProjectionPrivacyClassifier for SuspiciousWordClassifier {
-		fn classify_public_projection_text(
-			&self,
-			_field_name: &str,
-			text: &str,
-		) -> PublicProjectionPrivacyClassification {
-			if text.contains("private family detail") {
-				return PublicProjectionPrivacyClassification::Suspicious {
-					reason: String::from("matched fake private phrase"),
-				};
-			}
-
-			PublicProjectionPrivacyClassification::Allow
-		}
-	}
-
-	struct UnavailableClassifier;
-	impl PublicProjectionPrivacyClassifier for UnavailableClassifier {
-		fn classify_public_projection_text(
-			&self,
-			_field_name: &str,
-			_text: &str,
-		) -> PublicProjectionPrivacyClassification {
-			PublicProjectionPrivacyClassification::Unavailable {
-				reason: String::from("fake classifier unavailable"),
-			}
-		}
-	}
-
-	fn progress_record() -> LinearExecutionEventRecord {
-		LinearExecutionEventRecord::new(
-			LinearExecutionEventIdentity {
-				service_id: "decodex",
-				issue_id: "issue-id",
-				issue_identifier: "XY-519",
-				run_id: "xy-519-attempt-1",
-				attempt_number: 1,
-			},
-			"progress_checkpoint",
-			String::from("2026-05-25T00:00:00Z"),
-			"anchor",
-		)
-	}
-
-	#[test]
-	fn renders_attempt_number_as_run_sequence_in_comment_body() {
-		let record = LinearExecutionEventRecord::new(
-			LinearExecutionEventIdentity {
-				service_id: "decodex",
-				issue_id: "issue-id",
-				issue_identifier: "XY-519",
-				run_id: "xy-519-attempt-4",
-				attempt_number: 4,
-			},
-			"closeout",
-			String::from("2026-05-25T00:00:00Z"),
-			"anchor",
-		);
-		let body = super::render_linear_execution_event_comment_body(&record, Some(1));
-
-		assert!(body.contains("Decodex execution event: closeout"));
-		assert!(body.contains("- run_sequence_attempt: `4` (not retry-budget count)"));
-		assert!(body.contains("- retry_budget_attempts_consumed: `1`"));
-		assert!(!body.contains("- attempt:"));
-	}
-
-	#[test]
-	fn validates_public_progress_checkpoint_projection() {
-		let record = super::render_progress_checkpoint_public_projection(
-			LinearExecutionEventIdentity {
-				service_id: "decodex",
-				issue_id: "issue-id",
-				issue_identifier: "XY-519",
-				run_id: "xy-519-attempt-1",
-				attempt_number: 1,
-			},
-			String::from("2026-05-25T00:00:00Z"),
-			"implementing",
-			Some("y/decodex-xy-519"),
-			Some(".worktrees/XY-519"),
-			Some("https://github.com/hack-ink/decodex/pull/42"),
-		);
-
-		super::validate_linear_execution_event_record(&record)
-			.expect("public collaboration identifiers should validate");
-
-		assert_eq!(record.summary.as_deref(), Some("Execution phase: implementing."));
-		assert!(record.focus.is_none());
-		assert!(record.next_action.is_none());
-		assert!(record.evidence.is_none());
-	}
-
-	#[test]
-	fn rejects_private_progress_checkpoint_fields_in_linear_projection() {
-		for field_name in ["focus", "next_action", "blockers", "evidence", "verification"] {
-			let mut record = progress_record();
-
-			record.phase = Some(String::from("implementing"));
-			record.summary = Some(String::from("Execution phase: implementing."));
-
-			match field_name {
-				"focus" => record.focus = Some(String::from("private focus")),
-				"next_action" => record.next_action = Some(String::from("private action")),
-				"blockers" => record.blockers = Some(vec![String::from("private blocker")]),
-				"evidence" => record.evidence = Some(vec![String::from("private evidence")]),
-				"verification" => {
-					record.verification = Some(vec![String::from("private verification")]);
-				},
-				_ => unreachable!("test field names are exhaustive"),
-			}
-
-			let error = super::validate_linear_execution_event_record(&record)
-				.expect_err("private progress fields should not serialize");
-
-			assert!(error.contains("belongs in private execution events"));
-		}
-	}
-
-	#[test]
-	fn privacy_classifier_allows_public_projection_text() {
-		let mut record = progress_record();
-
-		record.phase = Some(String::from("implementing"));
-		record.summary = Some(String::from("Execution phase: implementing."));
-
-		let projection =
-			super::linear_execution_event_public_projection("", &record, &AllowingClassifier);
-
-		assert!(!projection.classifier_withheld_text);
-		assert_eq!(projection.record.summary.as_deref(), Some("Execution phase: implementing."));
-
-		super::validate_linear_execution_event_record(&projection.record)
-			.expect("allowed projection should remain valid");
-	}
-
-	#[test]
-	fn privacy_classifier_replaces_suspicious_required_summary_and_skips_optional_text() {
-		let mut record = progress_record();
-
-		record.phase = Some(String::from("implementing"));
-		record.summary = Some(String::from("Execution phase includes private family detail."));
-		record.raw_error = Some(String::from("raw private family detail"));
-
-		let projection = super::linear_execution_event_public_projection(
-			"comment has private family detail",
-			&record,
-			&SuspiciousWordClassifier,
-		);
-
-		assert!(projection.classifier_withheld_text);
-		assert_eq!(
-			projection.record.summary.as_deref(),
-			Some(super::PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_SUMMARY)
-		);
-		assert!(projection.record.raw_error.is_none());
-		assert_eq!(projection.body, super::PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_COMMENT_BODY);
-
-		super::validate_linear_execution_event_record(&projection.record)
-			.expect("classified projection should remain valid");
-	}
-
-	#[test]
-	fn unavailable_privacy_classifier_fails_closed_with_fixed_public_fields() {
-		let mut record = LinearExecutionEventRecord::new(
-			LinearExecutionEventIdentity {
-				service_id: "decodex",
-				issue_id: "issue-id",
-				issue_identifier: "XY-519",
-				run_id: "xy-519-attempt-1",
-				attempt_number: 1,
-			},
-			"terminal_failure",
-			String::from("2026-05-25T00:00:00Z"),
-			"anchor",
-		);
-
-		record.error_class = Some(String::from("repo_gate_failed"));
-		record.next_action = Some(String::from("inspect the failed command"));
-		record.blockers = Some(vec![String::from("repo gate failed")]);
-		record.evidence = Some(vec![String::from("cargo make test failed")]);
-
-		let projection =
-			super::linear_execution_event_public_projection("", &record, &UnavailableClassifier);
-
-		assert!(projection.classifier_withheld_text);
-		assert_eq!(
-			projection.record.next_action.as_deref(),
-			Some(super::PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_ACTION)
-		);
-		assert_eq!(
-			projection.record.blockers.as_deref(),
-			Some([String::from(super::PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_DETAIL)].as_slice())
-		);
-		assert_eq!(
-			projection.record.evidence.as_deref(),
-			Some([String::from(super::PRIVACY_CLASSIFIER_WITHHELD_PUBLIC_DETAIL)].as_slice())
-		);
-
-		super::validate_linear_execution_event_record(&projection.record)
-			.expect("unavailable classifier fallback should remain valid");
-	}
-}
+pub(crate) use self::types::{
+	CLOSEOUT_RECORD_TYPE, CloseoutRecord, REVIEW_HANDOFF_RECORD_TYPE, ReviewHandoffRecord,
+};
+pub(crate) use self::{
+	parse::{has_linear_execution_event_record, parse_linear_execution_event_record},
+	public_projection::linear_execution_event_public_projection,
+	render::{
+		append_structured_comment_record, render_linear_execution_event_comment_body,
+		render_progress_checkpoint_public_projection, stable_event_anchor,
+	},
+	types::{
+		LINEAR_EXECUTION_EVENT_RECORD_TYPE, LINEAR_EXECUTION_EVENT_RECORD_VERSION,
+		LinearExecutionEventIdentity, LinearExecutionEventPublicProjection,
+		LinearExecutionEventRecord,
+	},
+	validation::validate_linear_execution_event_record,
+};
